requirements.txt
altair
pandas
streamlit
plost
google-cloud-bigquery
google-api-core
google-api-python-client
altair
plotly
webcolors
matplotlib
pywaffle